Rename and fix forwarder deployment script
